fix: circles not showing on map
