+++ packages/backend/types/index.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 import {
   boundsSchema,
+  circleMeasurements,
   collection_linksEditSchema,
   collection_linksSchema,
   collectionsEditSchema,
@@ -9,6 +10,7 @@ import {
   combinedMarkersSchema,
   labelsEditSchema,
   labelsSchema,
+  lineMeasurements,
   mapsEditSchema,
   mapsSchema,
   mapTypeIdEnum,
@@ -16,11 +18,14 @@ import {
   mapUserSchema,
   markersEditSchema,
   markersSchema,
+  measurementsSchema,
   pathsEditSchema,
   pathsSchema,
   permissionEnumSchema,
   placesEditSchema,
   placesSchema,
+  polygonMeasurements,
+  rectangleMeasurements,
   refinedUserSchema,
   reviewsSchema,
   route_stopsEditSchema,
@@ -86,5 +91,11 @@ export type MapTypeId = z.infer<typeof mapTypeIdEnum>;
 
 export type Path = z.infer<typeof pathsSchema>;
 export type NewPath = z.infer<typeof pathsEditSchema>;
+
+export type Measurements = z.infer<typeof measurementsSchema>;
+export type CircleMeasurements = z.infer<typeof circleMeasurements>;
+export type RectangleMeasurements = z.infer<typeof rectangleMeasurements>;
+export type LineMeasurements = z.infer<typeof lineMeasurements>;
+export type PolygonMeasurements = z.infer<typeof polygonMeasurements>;
 
 export type PathStyle = z.infer<typeof stylesSchema>;
